ModuleProjects/ParkingArea/UsingPython/requirements.txt
# MySql Libraries
SQLAlchemy
PyMySQL